# backend-rust/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ABP Studio Rust API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/queue": {
      "get": {
        "operationId": "get_queue",
        "summary": "Migrated from Python get_queue",
        "responses": {
          "200": {
            "description": "Compatible JSON response"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "operationId": "add_to_queue",
        "summary": "Migrated from Python add_to_queue",
        "responses": {
          "200": {
            "description": "Compatible JSON response"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "delete": {
        "operationId": "clear_queue",
        "summary": "Migrated from Python clear_queue",
        "responses": {
          "200": {
            "description": "Compatible JSON response"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/queue/{item_id}": {
      "put": {
        "operationId": "update_queue_item",
        "summary": "Migrated from Python update_queue_item",
        "responses": {
          "200": {
            "description": "Compatible JSON response"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "delete": {
        "operationId": "delete_queue_item",
        "summary": "Migrated from Python delete_queue_item",
        "responses": {
          "200": {
            "description": "Compatible JSON response"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/v1/queue/{item_id}/retry": {
      "post": {
        "operationId": "retry_queue_item",
        "summary": "Migrated from Python retry_queue_item",
        "responses": {
          "200": {
            "description": "Compatible JSON response"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/v1/keywords/history": {
      "post": {
        "operationId": "save_keyword_history",
        "summary": "Migrated from Python save_keyword_history",
        "responses": {
          "200": {
            "description": "Compatible JSON response"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "operationId": "get_keyword_history",
        "summary": "Migrated from Python get_keyword_history",
        "responses": {
          "200": {
            "description": "Compatible JSON response"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "delete": {
        "operationId": "clear_keyword_history",
        "summary": "Migrated from Python clear_keyword_history",
        "responses": {
          "200": {
            "description": "Compatible JSON response"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/keywords/history/{index}": {
      "delete": {
        "operationId": "delete_keyword_history",
        "summary": "Migrated from Python delete_keyword_history",
        "responses": {
          "200": {
            "description": "Compatible JSON response"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  }
}